Extract CLI settings parsers from lib.rs into settings_parse
The crate root mixed the pure --preserve/--metadata-compare/--compare DSL parsers
(and the --update compare-vs-preserve validation) with the runtime/progress
machinery. Move them, with their two test modules, into a new settings_parse
module, re-exported from the root so common::parse_* is unchanged.

Also fix two pre-existing copy-paste error strings in parse_compare_settings
(cloned from parse_preserve_settings) that said "preserve settings" while
parsing the --compare DSL; flagged in review.
